requirements.txt
Django==1.5.4
GeoIP-Python==1.2.4
Jinja2==2.7.1
Mako==0.8.1
MarkupSafe==0.15
MySQL-python==1.2.3
PAM==0.4.2
PIL==1.1.7
Pillow==2.0.0
PyInstaller==2.1
Pygments==1.6
Scrapy==0.20.1
SecretStorage==1.0.0
South==0.8.4
Sphinx==1.1.3
Twisted==13.2.0
Twisted-Core==13.0.0
Twisted-Names==13.0.0
Twisted-Web==13.0.0
adium-theme-ubuntu==0.3.3
amqp==1.0.13
anyjson==0.3.3
apt-xapian-index==0.45
argparse==1.2.1
beautifulsoup4==4.3.1
billiard==2.7.3.31
bpython==0.12
celery==3.0.21
chardet==2.0.1
command-not-found==0.3
configglue==1.1.2
configobj==4.7.2
coverage==3.6
cssselect==0.9.1
debtagshw==0.1
decorator==3.4.0
defer==1.0.6
dirspec==13.10
distribute==0.6.37
django-allauth==0.15.0
django-braces==0.2.1
django-celery==3.0.17
django-choices==1.1.12
django-cors==0.1
django-cors-headers==0.12
django-crispy-forms==1.2.7
django-debug-toolbar==0.9.4
django-discover-runner==0.2.2
django-dynamic-scraper==0.3.5
django-endless-pagination==2.0
django-model-utils==1.1.0
django-storages==1.1.8
docutils==0.10
duplicity==0.6.21
facepy==0.8.4
gevent==1.0
greenlet==0.4.1
gunicorn==18.0
html5lib==0.95
httplib2==0.8
ipdb==0.8
ipython==0.13.2
joblib==0.7.0
jsonfield==0.9.17
keyring==1.6
kombu==2.5.16
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
librabbitmq==1.0.1
logutils==0.3.3
lxml==3.2.0
matplotlib==1.2.1
mechanize==0.2.5
meld3==0.6.5
mod-python==3.3.1
netifaces==0.8
nose==1.3.0
numexpr==2.0.1
numpy==1.7.1
oauth==1.0.1
oauthlib==0.5.1
obfsproxy==0.2.1
oneconf==0.3.5
openpyxl==1.5.8
pandas==0.12.0
pexpect==2.4
piston-mini-client==0.7.5
pkgconfig==1.1.0
protobuf==2.4.1
pyOpenSSL==0.13
pycrypto==2.6
pycups==1.9.63
pycurl==7.19.0
pygobject==3.10.0
pygpgme==0.3
pyinotify==0.9.3
pylibmc==1.2.3
pyparsing==2.0.1
pyptlib==0.0.3
pyserial==2.6
pysmbc==1.0.13
python-apt==0.8.9.1ubuntu1
python-dateutil==1.5
python-debian==0.1.21-nmu2ubuntu1
python-distutils-extra==2.38
python-openid==2.2.5
python-socksipy==1.0
python-tesseract==tesseract
pytz==2013b
pyxdg==0.25
pyzmq==13.1.0
queuelib==1.1.1
reportlab==2.6
requests==1.2.3
requests-oauthlib==0.3.2
rhythmbox-ubuntuone==13.10
scikit-learn==0.13.1
scipy==0.12.0
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
simplegeneric==0.8.1
six==1.4.1
software-center-aptd-plugins==0.0.0
sqlparse==0.1.10
ssh-import-id==3.19
statsmodels==0.4.2
stevedore==0.9.1
supervisor==3.0b2
sympy==0.7.2
system-service==0.1.6
tables==2.4.0
tornado==2.4.1
ubuntuone-storage-protocol==13.10
unity-lens-photos==1.0
urllib3==1.6
vboxapi==1.0
virtualenv==1.9.1
virtualenv-clone==0.2.4
virtualenvwrapper==4.0
w3lib==1.5
wadllib==1.3.2
wsgiref==0.1.2
wxPython==2.8.12.1
wxPython-common==2.8.12.1
xdiagnose==3.6.1
xkit==0.0.0
xlrd==0.9.2
xlwt==0.7.4
zope.interface==4.0.5
